+++ src/redux/selectors/index.js
@@ -1,10 +1,10 @@
 import { subjectsSelector } from './subjects';
 import { subjectSessionsSelector } from './workSessions';
-import { runningSessionSelector, selectedSubjectIdSelector } from './work';
+import { runningSessionSelector, selectedSubjectSelector } from './work';
 
 export {
   subjectsSelector,
   subjectSessionsSelector,
   runningSessionSelector,
-  selectedSubjectIdSelector,
+  selectedSubjectSelector,
 };