+++ src/redux/selectors/work.js
@@ -1,6 +1,7 @@
 import { createOrmSelector, ormSessionSelector } from './orm';
 
 const runningSessionIdSelector = state => state.work.runningSessionId;
+const selectedSubjectIdSelector = state => state.work.selectedSubjectId;
 
 export const runningSessionSelector = createOrmSelector(
   ormSessionSelector,
@@ -9,4 +10,9 @@ export const runningSessionSelector = createOrmSelector(
     && ormSession.WorkSession.withId(runningSessionId),
 );
 
-export const selectedSubjectIdSelector = state => state.work.selectedSubjectId;
+export const selectedSubjectSelector = createOrmSelector(
+  ormSessionSelector,
+  selectedSubjectIdSelector,
+  (ormSession, selectedSubjectId) => selectedSubjectId !== -1
+    && ormSession.Subject.withId(selectedSubjectId),
+);