+++ docs/CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Upcoming
+### Added
+* Icon and buttons in notifications
+
 ## Unreleased
 ### Added
 * Settings menu, with version number
+* Notification to show the user if is playing or paused. Display only title
 
 ## 0.0.2 - 2018-08-23
 ### Fixed